Move units around.
