+++ source/Quantities.Units/Si/Metric/Gram.cs
@@ -1,10 +1,10 @@
 ﻿using Quantities.Dimensions;
 
-namespace Quantities.Units.Si.Derived;
+namespace Quantities.Units.Si.Metric;
 
 public readonly struct Gram : IMetricUnit, IMass
 {
-    public static Transformation ToSi(Transformation self) => self.RootedIn<Metre>() / 1000;
+    public static Transformation ToSi(Transformation self) => self.RootedIn<Kilogram>() / 1000;
     public static String Representation => "g";
 
 }
